feat(chat): add console, preview, and code tabs with responsive toolbar and save changes functionality
