+++ frontend/src/components/chat/code-engine/code-engine.tsx
@@ -1,24 +1,14 @@
 'use client';
-import { Button } from '@/components/ui/button';
-import Editor from '@monaco-editor/react';
-import { ExclamationTriangleIcon } from '@radix-ui/react-icons';
-import { motion, AnimatePresence } from 'framer-motion';
-import {
-  Code as CodeIcon,
-  Copy,
-  Eye,
-  GitFork,
-  Loader,
-  Share2,
-  Terminal,
-} from 'lucide-react';
-import { useTheme } from 'next-themes';
 import { useContext, useEffect, useRef, useState } from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+import { Loader } from 'lucide-react';
 import { TreeItem, TreeItemIndex } from 'react-complex-tree';
-import FileExplorerButton from './file-explorer-button';
-import FileStructure from './file-structure';
 import { ProjectContext } from './project-context';
-import WebPreview from './web-view';
+import CodeTab from './tabs/code-tab';
+import PreviewTab from './tabs/preview-tab';
+import ConsoleTab from './tabs/preview-tab';
+import ResponsiveToolbar from './responsive-toolbar';
+import SaveChangesBar from './save-changes-bar';
 
 export function CodeEngine({
   chatId,
@@ -29,80 +19,57 @@ export function CodeEngine({
   isProjectReady?: boolean;
   projectId?: string;
 }) {
-  // Initialize state, refs, and context
-  const editorRef = useRef(null);
+  // Initialize state and context
   const { curProject, filePath } = useContext(ProjectContext);
+  console.log('filepath', filePath);
+  console.log('chatid', chatId);
+  console.log('projectid', projectId);
+  const [saving, setSaving] = useState(false);
   const [preCode, setPrecode] = useState('// some comment');
   const [newCode, setCode] = useState('// some comment');
-  const [saving, setSaving] = useState(false);
-  const [type, setType] = useState('javascript');
-  const [isLoading, setIsLoading] = useState(false);
-  const [isExplorerCollapsed, setIsExplorerCollapsed] = useState(false);
-  const [fileStructureData, setFileStructureData] = useState<
-    Record<TreeItemIndex, TreeItem<any>>
-  >({});
-  console.log(fileStructureData);
-  const theme = useTheme();
-
   const [activeTab, setActiveTab] = useState<'preview' | 'code' | 'console'>(
     'code'
   );
+  const [isFileStructureLoading, setIsFileStructureLoading] = useState(false);
+  const [fileStructureData, setFileStructureData] = useState<
+    Record<TreeItemIndex, TreeItem<any>>
+  >({});
 
-  // Callback: Handle editor mount
-  const handleEditorMount = (editorInstance) => {
-    editorRef.current = editorInstance;
-    // Set the editor DOM node's position for layout control
-    editorInstance.getDomNode().style.position = 'absolute';
-  };
+  // Reference to the editor instance
+  const editorRef = useRef(null);
 
-  // Effect: Fetch file content when filePath or projectId changes
-  useEffect(() => {
-    async function getCode() {
-      if (!curProject || !filePath) return;
-
-      const file_node = fileStructureData[`root/${filePath}`];
-      if (filePath == '' || !file_node) return;
-      const isFolder = file_node.isFolder;
-      if (isFolder) return;
-      try {
-        setIsLoading(true);
-        const res = await fetch(
-          `/api/file?path=${encodeURIComponent(`${curProject.projectPath}/${filePath}`)}`
-        ).then((res) => res.json());
-        setCode(res.content);
-        setPrecode(res.content);
-        setType(res.type);
-        setIsLoading(false);
-      } catch (error: any) {
-        console.error(error.message);
-      }
+  // Fetch file structure function (only for code tab)
+  async function fetchFiles() {
+    if (!curProject?.projectPath) {
+      return;
     }
-    getCode();
-  }, [filePath, curProject, fileStructureData]);
 
-  // Effect: Fetch file structure when projectId changes
-  useEffect(() => {
-    async function fetchFiles() {
-      if (!curProject?.projectPath) {
-        console.log('no project path found');
-        return;
-      }
-
-      try {
-        const response = await fetch(
-          `/api/project?path=${curProject.projectPath}`
-        );
-        console.log('loading file structure');
-        const data = await response.json();
-        setFileStructureData(data.res || {});
-      } catch (error) {
-        console.error('Error fetching file structure:', error);
-      }
+    try {
+      setIsFileStructureLoading(true);
+      const response = await fetch(
+        `/api/project?path=${curProject.projectPath}`
+      );
+      const data = await response.json();
+      setFileStructureData(data.res || {});
+    } catch (error) {
+      console.error('Error fetching file structure:', error);
+    } finally {
+      setIsFileStructureLoading(false);
     }
-    fetchFiles();
-  }, [curProject?.projectPath]);
+  }
 
-  // Reset code to previous state and update editor
+  // Effect: Fetch file structure when needed for code tab
+  useEffect(() => {
+    if (
+      activeTab === 'code' &&
+      Object.keys(fileStructureData).length === 0 &&
+      curProject?.projectPath
+    ) {
+      fetchFiles();
+    }
+  }, [activeTab, fileStructureData, curProject?.projectPath]);
+
+  // Reset code to previous state
   const handleReset = () => {
     setCode(preCode);
     editorRef.current?.setValue(preCode);
@@ -136,148 +103,43 @@ export function CodeEngine({
     updateCode(newCode);
   };
 
-  // Update code in state and mark as saving
+  // Track unsaved changes
   const updateSavingStatus = (value) => {
     setCode(value);
     setSaving(true);
   };
 
-  // Responsive toolbar component for header tabs and buttons
-  const ResponsiveToolbar = ({ isLoading }: { isLoading: boolean }) => {
-    const containerRef = useRef(null);
-    const [containerWidth, setContainerWidth] = useState(700);
-    const [visibleTabs, setVisibleTabs] = useState(3);
-    const [compactIcons, setCompactIcons] = useState(false);
-
-    // Observe container width changes
-    useEffect(() => {
-      const observer = new ResizeObserver((entries) => {
-        for (const entry of entries) {
-          setContainerWidth(entry.contentRect.width);
-        }
-      });
-
-      if (containerRef.current) {
-        observer.observe(containerRef.current);
-      }
-      return () => observer.disconnect();
-    }, []);
-
-    // Adjust visible tabs and icon style based on container width
-    useEffect(() => {
-      if (containerWidth > 650) {
-        setVisibleTabs(3);
-        setCompactIcons(false);
-      } else if (containerWidth > 550) {
-        setVisibleTabs(2);
-        setCompactIcons(false);
-      } else if (containerWidth > 450) {
-        setVisibleTabs(1);
-        setCompactIcons(true);
-      } else {
-        setVisibleTabs(0);
-        setCompactIcons(true);
-      }
-    }, [containerWidth]);
-
-    return (
-      <div
-        ref={containerRef}
-        className="flex items-center justify-between p-4 border-b w-full bg-background/60 backdrop-blur supports-[backdrop-filter]:bg-background/60"
-      >
-        <div className="flex items-center space-x-2">
-          <Button
-            variant={activeTab === 'preview' ? 'default' : 'outline'}
-            className="text-sm"
-            onClick={() => setActiveTab('preview')}
-            disabled={isLoading}
-          >
-            <Eye className="w-4 h-4 mr-1" />
-            Preview
-          </Button>
-          {visibleTabs >= 2 && (
-            <Button
-              variant={activeTab === 'code' ? 'default' : 'outline'}
-              className="text-sm"
-              onClick={() => setActiveTab('code')}
-              disabled={isLoading}
-            >
-              <CodeIcon className="w-4 h-4 mr-1" />
-              Code
-            </Button>
-          )}
-          {visibleTabs >= 3 && (
-            <Button
-              variant={activeTab === 'console' ? 'default' : 'outline'}
-              className="text-sm"
-              onClick={() => setActiveTab('console')}
-              disabled={isLoading}
-            >
-              <Terminal className="w-4 h-4 mr-1" />
-              Console
-            </Button>
-          )}
-        </div>
-
-        <div className="flex items-center space-x-2">
-          <div className="flex items-center space-x-2">
-            <Button
-              variant="ghost"
-              className={`p-0 ${compactIcons ? 'hidden' : 'block'}`}
-              disabled={isLoading}
-            >
-              <GitFork className="w-5 h-5" />
-            </Button>
-            <Button
-              variant="ghost"
-              className={`p-0 ${compactIcons ? 'hidden' : 'block'}`}
-              disabled={isLoading}
-            >
-              <Share2 className="w-5 h-5" />
-            </Button>
-            <Button
-              variant="ghost"
-              className={`p-0 ${compactIcons ? 'hidden' : 'block'}`}
-              disabled={isLoading}
-            >
-              <Copy className="w-5 h-5" />
-            </Button>
-          </div>
-          <div className="flex items-center space-x-2">
-            {!compactIcons && (
-              <>
-                <Button
-                  variant="outline"
-                  className="text-sm"
-                  disabled={isLoading}
-                >
-                  Supabase
-                </Button>
-                <Button
-                  variant="outline"
-                  className="text-sm"
-                  disabled={isLoading}
-                >
-                  Publish
-                </Button>
-              </>
-            )}
-            {compactIcons && (
-              <Button variant="outline" className="p-2" disabled={isLoading}>
-                <Share2 className="w-4 h-4" />
-              </Button>
-            )}
-          </div>
-        </div>
-      </div>
-    );
+  // Render appropriate content based on active tab
+  const renderTabContent = () => {
+    switch (activeTab) {
+      case 'code':
+        return (
+          <CodeTab
+            editorRef={editorRef}
+            fileStructureData={fileStructureData}
+            newCode={newCode}
+            isFileStructureLoading={isFileStructureLoading}
+            updateSavingStatus={updateSavingStatus}
+          />
+        );
+      case 'preview':
+        return <PreviewTab />;
+      case 'console':
+        return <ConsoleTab />;
+      default:
+        return null;
+    }
   };
 
   // Render the CodeEngine layout
   return (
     <div className="rounded-lg border shadow-sm overflow-scroll h-full">
       {/* Header Bar */}
-      <ResponsiveToolbar isLoading={!isProjectReady} />
+      <ResponsiveToolbar
+        isLoading={!isProjectReady}
+        activeTab={activeTab}
+        setActiveTab={setActiveTab}
+      />
 
       {/* Main Content Area with Loading */}
       <div className="relative h-[calc(100vh-48px-4rem)]">
@@ -298,99 +160,13 @@ export function CodeEngine({
           )}
         </AnimatePresence>
 
-        <div className="flex h-full">
-          {activeTab === 'code' ? (
-            <>
-              {/* File Explorer Panel (collapsible) */}
-              <motion.div
-                animate={{
-                  width: isExplorerCollapsed ? '0px' : '300px',
-                  opacity: isExplorerCollapsed ? 0 : 1,
-                }}
-                transition={{ duration: 0.3, ease: 'easeInOut' }}
-                className="overflow-y-auto border-r"
-              >
-                <FileStructure data={fileStructureData} filePath={filePath} />
-              </motion.div>
-              <div className="flex-1 relative">
-                <Editor
-                  height="100%"
-                  width="100%"
-                  defaultLanguage="typescriptreact"
-                  value={newCode}
-                  language={type}
-                  loading={isLoading}
-                  onChange={updateSavingStatus}
-                  onMount={handleEditorMount}
-                  options={{
-                    fontSize: 14,
-                    minimap: { enabled: false },
-                    wordWrap: 'on',
-                    wrappingStrategy: 'advanced',
-                    scrollbar: {
-                      useShadows: false,
-                      vertical: 'visible',
-                      horizontal: 'visible',
-                      verticalScrollbarSize: 10,
-                      horizontalScrollbarSize: 10,
-                    },
-                  }}
-                  theme={theme.theme === 'dark' ? 'vs-dark' : 'vs'}
-                />
-              </div>
-            </>
-          ) : activeTab === 'preview' ? (
-            <div className="w-full h-full">
-              <WebPreview />
-            </div>
-          ) : activeTab === 'console' ? (
-            <div className="flex-1 p-4 text-sm">Console Content (Mock)</div>
-          ) : null}
-        </div>
+        <div className="flex h-full">{renderTabContent()}</div>
 
         {/* Save Changes Bar */}
-        {saving && (
-          <SaveChangesBar
-            saving={saving}
-            onSave={handleSave}
-            onReset={handleReset}
-          />
-        )}
-
-        {/* File Explorer Toggle Button */}
-        {activeTab === 'code' && (
-          <FileExplorerButton
-            isExplorerCollapsed={isExplorerCollapsed}
-            setIsExplorerCollapsed={setIsExplorerCollapsed}
-          />
-        )}
+        {saving && <SaveChangesBar onSave={handleSave} onReset={handleReset} />}
       </div>
     </div>
   );
 }
 
-// SaveChangesBar component for showing unsaved changes status
-const SaveChangesBar = ({ saving, onSave, onReset }) => {
-  return (
-    saving && (
-      <div className="fixed bottom-4 right-4 flex items-center space-x-2 p-2 border rounded-full shadow bg-background">
-        <ExclamationTriangleIcon className="w-5 h-5 text-yellow-500" />
-        <span className="text-sm text-foreground">Unsaved Changes</span>
-        <Button
-          variant="outline"
-          className="px-3 py-1 text-sm font-medium rounded-full"
-          onClick={onReset}
-        >
-          Reset
-        </Button>
-        <Button
-          variant="default"
-          className="px-4 py-1 text-sm font-medium rounded-full"
-          onClick={onSave}
-        >
-          Save
-        </Button>
-      </div>
-    )
-  );
-};
+export default CodeEngine;
