+++ frontend/src/components/chat/code-engine/tabs/preview-tab.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+const ConsoleTab = () => {
+  return <div className="flex-1 p-4 text-sm">Console Content (Mock)</div>;
+};
+
+export default ConsoleTab;
